Refined toString matching
This test case failed on Travis but not local. I forgot to add the new isValid field to the matching. Because the regex took everything at the end of the toString block, it was still valid locally because it was the last value. The JVM Travis uses seems to create a different ordering, making it the first entry.

Now the test just asks if the result resembles mainly the ByteBuddy toString() method, not the printed parameter.