Added Domain Configuration concept
Its now possible to create a domain configuration. This allows to define a special logic for field/class setting. This could be that a String field called `surname` gets a static value or is directed to the Faker library to get something that is semantically a surname. Similar is possible for fields like `pin`, `uuid` or in general fields that are used frequently in a project.