Best Match Factory finder
The #createRandomFilledByFactory method without the additional MethodHolder will try to identify a static method which returns the desired class and does not have the desired class as parameter. Later should omit copy factory methods which would end in possible NPEs since Fakegen would try to create the class directly, but cannot, and return null for the parameter. :bangbang: