Refactored Domain Handling
The previous solution had two different approaches. A filler class with provided generation of specific objects, currently just the java basic classes. The other is the domain handling which allows to to define a specific behaviour if the field has a certain name. Problematic on this approach was the growing amount of different parameters and how they depend on each other. It was not elegantly possible to allow the domain handling to use the basic generation again.