Test Cases for Junit 5
To test the junit 5 extension, I needed the code the Junit ppl use to test their framework. This code is located within their test space and therefore not directly available as dependency. With the new 5.4 Release a seperate `junit-platform-testkit` will be released, but for now I adapted their code. It is explicitly marked as under EPL 2.0.