#31 Added support for Optional with simple types
ObjectFiller methods are now required to to accept also the `Field`. With the field we can infer the generic type, and later on annotations on the field. Currently it has the limitation that only simple type parameters are supported. An `Optional<List<String>>` would not be possible :bangbang:. Later on a new internal creation routine is required within works with a `ParameterizedType`.