#31 Added support for Optional with ParameterizedTypes
To fill an Optional we need to get the actual type Arguments presented. In a simple situation this can be a regular Class (String, MyClass, etc). In complex Situation the Type presented to the Optional is again a Parameterized Type (List<String>, Map<String, String>, etc.) The new fillInstance method for Types creates a new instance of the raw type (here List or Map).