#10 Added Configuration object
Added Config object which currently allows defining if TestDataFiller should use private Constructors, setters over field Injection and a set of fields to ignore. In the future, we could allow a deeper recursion or the usage of a cache.

Additionally added two default configurations. Notable is the BUILDER config. It excludes the field `initBits`, because it is used by the Immutables Builder to track the completion of required fields.